nvmesensor: Add toggle for single thread mode
Add a flag to NVMeMi class to turn off the single thread mode. The
single thread mode is to aggregate the communicate to the devices under
the same i2c bus. So the MCTP request/response packet will always
transmitting in the expected order.

If the I2C drive support concurrent master/slave mode, the single thread
mode is not required since the MCTP layer can handle the concurrent MCTP
packets to/from different endpoints. Then the toggle can be turned off.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: Ib17650c9d6465167f046dee44e2d526255343f1b