nvmesensor: add PowerState to NVMeMi
Some NVMe vendor implementation link the OOB NVMe-MI connectivity to the
inband signal. For example, there will be MCTP level reset when inband
PERST is issued.

The proposed solution is to NVMeMi will close the MCTP ep upon
PowerState down, and re-init the MCTP when PowerState comes back up.

Configuration for PowerState of each NVMeMi MCTP endpoint is via E-M
configuration of "PowerState". The default value is always, and the
configu value is case insensitive.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I7e4c84415fb3568226bc51a03279aeec32a5c612