nvmesensor: improve handling of config change
The improvement includes:
1. Make sure there is only one `handleConfigurations` function running
globally. That function is scheduled asynchroneously to read the
config profiles from dbus. Given the reading delay is uncertain and
could be possibly longer than the filter timer for property change
events, there could be multiple `handleConfigurations` function being
scheduling and stack in the asio queue. In this case, we only need to
handle the latest config profile reading instead of all of them.

2. Move the subsystem clean-up process ahead of the config profile
reading and make a miminal interval of 5 seconds between the
subsystem destroy and re-creation. That is because the lifetime the nvme
subsystem is held asynchroneously by its internal polling loop. Stop the
subsystem only sends a signal to break the loop but the actual resource
recollection is not completed until all related asio context has been
safely drained from the io queue. 5 seconds is a practise value based on
1 second of polling rate and communication delay < 1s.

Signed-off-by: Hao Jiang <jianghao@google.com>
Change-Id: I5d029cef792806f0f2861ad03125acd63b6dc3c7